refactor: 패키지 역할 분리
- domain과 repository의 의미를 명확히 분리
- domain은 동전 하나, 상품 하나 등 어떤 객체의 틀
- repository는 여러 개의 domain 객체를 가지는 저장소의 의미
