+++ src/main/java/vendingmachine/repository/Change.java
@@ -1,7 +1,9 @@
-package vendingmachine.domain;
+package vendingmachine.repository;
 
 import java.util.HashMap;
 import java.util.Map;
+
+import vendingmachine.domain.Coin;
 
 public class Change {
